Add flip parameter to `geom_violinhalf()`
This commit adds an extra parameter `flip` that can be passed into the violinhalf geom. By default it is `FALSE` and the flat-side of the violin will be on the left; if `flip = TRUE`, then the direction of all half violins will switch and the flat-side will be on the right.