Robust patchwork printing
Make check_model and visualisation_recipe plotting robust to {grid} viewport errors when the RStudio Plots pane is too small.

We can't do anything about making the plot show, but we can give a more useful error message. I submitted a PR to {patchwork}, but we can use it here in the meantime.

@IndrajeetPatil could be useful for {modelsummary}